v0.9.0 - Overhaul Pipelines
