+++ harvey/__init__.py
@@ -1,10 +1,10 @@
 # flake8: noqa
-from harvey.globals import Global
-from harvey.container import Container
+from harvey.containers import Container
 from harvey.git import Git
-from harvey.image import Image
-from harvey.message import Message
-from harvey.pipeline import Pipeline
-from harvey.stage import Stage
-from harvey.utils import Utils, Logs
-from harvey.webhook import Webhook
+from harvey.globals import Global
+from harvey.images import Image
+from harvey.messages import Message
+from harvey.pipelines import Pipeline
+from harvey.stages import Stage
+from harvey.utils import Logs, Utils
+from harvey.webhooks import Webhook
