+++ harvey/images.py
