+++ test/unit/test_app.py
@@ -16,14 +16,14 @@ def test_routes_are_reachable_get(mock_client, route):
     assert response.status_code == 200
 
 
-@mock.patch('harvey.webhook.Webhook.parse_webhook')
+@mock.patch('harvey.webhooks.Webhook.parse_webhook')
 def test_start_pipeline(mock_parse_webhook):
     # TODO: Long-term, test the status_code and logic
     app.start_pipeline()
     mock_parse_webhook.assert_called_once()
 
 
-@mock.patch('harvey.webhook.Webhook.parse_webhook')
+@mock.patch('harvey.webhooks.Webhook.parse_webhook')
 def test_start_pipeline_compose(mock_parse_webhook):
     # TODO: Long-term, test the status_code and logic
     app.start_pipeline_compose()
